clean up and poststatus
clean up and poststatus

+++ app/src/main/java/com/dnjagi/carval/MainActivity.java
@@ -7,13 +7,11 @@ import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
+import android.os.Handler;
 import android.support.annotation.RequiresApi;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentTransaction;
-import android.view.View;
 import android.support.design.widget.NavigationView;
 import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
@@ -22,12 +20,17 @@ import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
+import android.widget.Toast;
 
 import com.dnjagi.carval.data.LogRecord;
-import com.dnjagi.carval.data.SuspensionSystemRecord;
 import com.dnjagi.carval.database.SchemaGenerator;
 import com.dnjagi.carval.database.SugarContext;
 import com.dnjagi.carval.database.SugarDb;
+
+import java.net.URL;
+import java.util.Enumeration;
+
+import Services.UploadService;
 
 public class MainActivity extends AppCompatActivity
         implements NavigationView.OnNavigationItemSelectedListener,
@@ -42,17 +45,17 @@ public class MainActivity extends AppCompatActivity
         AccidentAssessmentFragment.OnFragmentInteractionListener {
 
     public static Context appContext;
+    private boolean cleanUpDatabase = false;
+
     @RequiresApi(api = Build.VERSION_CODES.M)
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
-
-
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
-
+        //request for permissions
+        GrantPermissions();
         DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
         ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
                 this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
@@ -62,44 +65,42 @@ public class MainActivity extends AppCompatActivity
         NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
         navigationView.setNavigationItemSelectedListener(this);
 
-      boolean y =  isStoragePermissionGranted();
-
-
-        appContext = getApplicationContext();
-        SugarContext.init(getApplicationContext());
-
+        appContext = this.getApplicationContext();
+        SugarContext.init(this.getApplicationContext());
         SchemaGenerator schemaGenerator = new SchemaGenerator(getApplicationContext());
-        schemaGenerator.deleteTables(new SugarDb(getApplicationContext()).getDB());
-        SugarContext.init(getApplicationContext());
+        boolean tableExists = schemaGenerator.tableExists(LogRecord.class, new SugarDb(getApplicationContext()).getDB());
+        if (!tableExists) {
+            schemaGenerator.deleteTables(new SugarDb(getApplicationContext()).getDB());
+            SugarContext.init(getApplicationContext());
+            schemaGenerator.createDatabase(new SugarDb(getApplicationContext()).getDB());
 
-        schemaGenerator.createDatabase(new SugarDb(getApplicationContext()).getDB());
+        }
 
-        boolean isTableCreated = schemaGenerator.tableExists(LogRecord.class, new SugarDb(getApplicationContext()).getDB());
-        if (!isTableCreated) {
+        //when need to clean up database
+        if (cleanUpDatabase) {
+            schemaGenerator.deleteTables(new SugarDb(getApplicationContext()).getDB());
+            SugarContext.init(getApplicationContext());
             schemaGenerator.createDatabase(new SugarDb(getApplicationContext()).getDB());
         }
+
+        //Start background services
+        Intent intent = new Intent(MainActivity.this, UploadService.class);
+        startService(intent);
 
         // go to home page
         FragmentTransaction ft = getSupportFragmentManager().beginTransaction();
         ft.replace(R.id.mainFrame, new HomeFragment());
         ft.commit();
-
-
     }
 
 
-    public boolean isStoragePermissionGranted() {
+    public void GrantPermissions() {
         if (Build.VERSION.SDK_INT >= 23) {
-            if (checkSelfPermission(android.Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED && checkSelfPermission(Manifest.permission.READ_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED
-                    ) {
-                return true;
+            if (checkSelfPermission(android.Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED &&
+                    checkSelfPermission(Manifest.permission.READ_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
             } else {
-
                 ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE, Manifest.permission.READ_EXTERNAL_STORAGE}, 1);
-                return false;
             }
-        } else { //permission is automatically granted on sdk<23 upon installation
-            return true;
         }
     }
 
@@ -108,7 +109,7 @@ public class MainActivity extends AppCompatActivity
         super.onRequestPermissionsResult(requestCode, permissions, grantResults);
         if (grantResults.length > 0) {
             if (grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                // LogSession();
+                Toast toast = Toast.makeText(getApplicationContext(), "Welcome" , Toast.LENGTH_LONG);
             }
         }
     }
@@ -137,12 +138,10 @@ public class MainActivity extends AppCompatActivity
         // automatically handle clicks on the Home/Up button, so long
         // as you specify a parent activity in AndroidManifest.xml.
         int id = item.getItemId();
-
         //noinspection SimplifiableIfStatement
         if (id == R.id.action_settings) {
             return true;
         }
-
         return super.onOptionsItemSelected(item);
     }
 
@@ -150,17 +149,13 @@ public class MainActivity extends AppCompatActivity
     @SuppressWarnings("StatementWithEmptyBody")
     @Override
     public boolean onNavigationItemSelected(MenuItem item) {
-
         // Handle navigation view item clicks here.
         int id = item.getItemId();
-
         //NOTE: creating fragment object
         Fragment fragment = null;
-
         if (id == R.id.car_valuation) {
             fragment = new CarValuationFragment();
         } else if (id == R.id.accident_assessment) {
-
             fragment = new AccidentAssessmentFragment();
         } else if (id == R.id.reports) {
             fragment = new ReportsFragment();
@@ -169,7 +164,6 @@ public class MainActivity extends AppCompatActivity
             Intent intent = new Intent(MainActivity.this, LoginActivity.class);
             startActivity(intent);
         }*/
-
 
         final Fragment finalFragment = fragment;
         new Thread(new Runnable() {
@@ -186,10 +180,8 @@ public class MainActivity extends AppCompatActivity
                 runOnUiThread(new Runnable() {
                     @Override
                     public void run() {
-
                         DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout); //Ya you can also globalize this variable :P
                         drawer.closeDrawer(GravityCompat.START);
-
                     }
                 });
             }
@@ -200,8 +192,5 @@ public class MainActivity extends AppCompatActivity
 
     @Override
     public void onFragmentInteraction(Uri uri) {
-
     }
-
-
 }
